React view for external-component-loader
