+++ components/experimental/external-component-loader/package.json
@@ -42,6 +42,8 @@
     "@microsoft/fluid-runtime-definitions": "^0.18.0",
     "@microsoft/fluid-view-interfaces": "^0.18.0",
     "@microsoft/fluid-web-code-loader": "^0.18.0",
+    "react": "^16.10.2",
+    "react-dom": "^16.10.2",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
@@ -53,6 +55,8 @@
     "@types/jest-environment-puppeteer": "2.2.0",
     "@types/node": "^10.14.6",
     "@types/puppeteer": "1.3.0",
+    "@types/react": "^16.9.15",
+    "@types/react-dom": "^16.9.4",
     "@typescript-eslint/eslint-plugin": "~2.17.0",
     "@typescript-eslint/eslint-plugin-tslint": "~2.16.0",
     "@typescript-eslint/parser": "~2.17.0",
